feat: create CJS and ESM bundles using rollup
