+++ packages/styled-system/package.json
@@ -2,20 +2,21 @@
   "name": "@tonic-ui/styled-system",
   "version": "1.2.2",
   "description": "The framework agnostic styling engine for Tonic UI.",
-  "main": "index.js",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.esm.js",
   "publishConfig": {
     "access": "public",
     "directory": "build"
   },
   "scripts": {
-    "build": "babel --root-mode upward src --out-dir build --ignore='**/__tests__/*'",
-    "clean": "del build coverage",
+    "build": "cross-env PACKAGE_NAME=index INPUT=src OUTPUT_DIRECTORY=dist rollup --config rollup.config.mjs",
+    "clean": "del build coverage dist",
     "eslint": "eslint --ext .js,.jsx,.mjs .",
     "package-name": "echo -e \"\\e[1;33m$npm_package_name\\e[0m\"",
-    "postbuild": "cpy {LICENSE,README.md,.npmignore,package.json} build",
+    "postbuild": "cpy {dist,LICENSE,README.md,.npmignore,package.json} build",
     "pre-push": "yarn --silent run package-name && yarn run build && yarn run eslint && yarn run test",
     "prepublish": "yarn run build",
-    "test": "jest"
+    "test": "jest --maxWorkers=2"
   },
   "devDependencies": {
     "@babel/cli": "^7.0.0",
@@ -24,6 +25,7 @@
     "@babel/eslint-plugin": "^7.0.0",
     "@babel/plugin-transform-runtime": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.0",
     "@trendmicro/babel-config": "^1.0.2",
     "cpy-cli": "^4.2.0",
     "cross-env": "^7.0.3",
@@ -35,7 +37,9 @@
     "eslint-plugin-react": "latest",
     "eslint-plugin-react-hooks": "latest",
     "jest": "^29.0.0",
-    "jest-environment-node": "^29.0.0"
+    "jest-environment-node": "^29.0.0",
+    "rollup": "^3",
+    "rollup-plugin-babel": "^4.4.0"
   },
   "jest": {
     "collectCoverage": true,
