+++ packages/styled-system/src/__tests__/index.test.js
@@ -3,9 +3,14 @@ import * as moduleExport from '..';
 test('should match expected exports', () => {
   const receivedExports = Object.keys(moduleExport);
   const expectedExports = [
+    // pseudo
     'pseudoClassSelector',
     'pseudoElementSelector',
+
+    // sx
     'sx',
+
+    // system
     'system',
   ];
