+++ packages/theme/src/__tests__/createTheme.test.js
@@ -1,4 +1,265 @@
-import rem from '../rem';
+import createTheme from '../createTheme';
+
+test('absolute length units: px', () => {
+  const {
+    borders,
+    breakpoints,
+    colors,
+    fonts,
+    fontSizes,
+    fontWeights,
+    letterSpacings,
+    lineHeights,
+    outlines,
+    radii,
+    shadows,
+    sizes,
+    space,
+    zIndices,
+  } = createTheme('px');
+
+  expect(borders).toEqual({
+    1: '1px solid',
+    2: '2px solid',
+    3: '3px solid',
+    4: '4px solid',
+    5: '5px solid',
+  });
+
+  expect([...breakpoints]).toEqual([
+    '320px',
+    '640px',
+    '1024px',
+    '1280px',
+    '1680px',
+  ]);
+
+  expect(colors).toEqual({
+    transparent: 'transparent',
+    current: 'currentColor',
+    'red:100': '#6e0002',
+    'red:90': '#9d0003',
+    'red:80': '#b80003',
+    'red:70': '#d71920',
+    'red:60': '#e52630',
+    'red:50': '#f24c4f',
+    'red:40': '#f46f71',
+    'red:30': '#fd999a',
+    'red:20': '#fcc3c4',
+    'red:10': '#fee1e2',
+    'magenta:100': '#750037',
+    'magenta:90': '#960043',
+    'magenta:80': '#b3004c',
+    'magenta:70': '#ca0455',
+    'magenta:60': '#dc1d68',
+    'magenta:50': '#e94181',
+    'magenta:40': '#f36fa0',
+    'magenta:30': '#f9a0c1',
+    'magenta:20': '#fcc3d8',
+    'magenta:10': '#fee1ec',
+    'purple:100': '#460086',
+    'purple:90': '#5300a5',
+    'purple:80': '#6304ca',
+    'purple:70': '#771ddc',
+    'purple:60': '#8f41e9',
+    'purple:50': '#ab6ff3',
+    'purple:40': '#bb89f6',
+    'purple:30': '#cca6f9',
+    'purple:20': '#ddc3fc',
+    'purple:10': '#eee1fe',
+    'blue:100': '#002a7e',
+    'blue:90': '#00349d',
+    'blue:80': '#003db8',
+    'blue:70': '#0547cd',
+    'blue:60': '#1e5ede',
+    'blue:50': '#578aef',
+    'blue:40': '#6f9bf4',
+    'blue:30': '#95b7fc',
+    'blue:20': '#c3d6fc',
+    'blue:10': '#e1ebfe',
+    'green:100': '#003011',
+    'green:90': '#00461a',
+    'green:80': '#005c24',
+    'green:70': '#00712e',
+    'green:60': '#008539',
+    'green:50': '#00a94f',
+    'green:40': '#04c45a',
+    'green:30': '#40e884',
+    'green:20': '#89f6b2',
+    'green:10': '#c3fcd8',
+    'teal:100': '#004034',
+    'teal:90': '#005242',
+    'teal:80': '#006451',
+    'teal:70': '#00755f',
+    'teal:60': '#00866c',
+    'teal:50': '#00a584',
+    'teal:40': '#04caa1',
+    'teal:30': '#41e9c5',
+    'teal:20': '#89f6df',
+    'teal:10': '#c3fcf0',
+    'cyan:100': '#003664',
+    'cyan:90': '#004575',
+    'cyan:80': '#005486',
+    'cyan:70': '#006496',
+    'cyan:60': '#0075a5',
+    'cyan:50': '#0095bf',
+    'cyan:40': '#10b4d3',
+    'cyan:30': '#41d8e9',
+    'cyan:20': '#89f0f6',
+    'cyan:10': '#c3f9fc',
+    'gray:100': '#151515',
+    'gray:90': '#212121',
+    'gray:80': '#303030',
+    'gray:70': '#424242',
+    'gray:60': '#5e5e5e',
+    'gray:50': '#8a8a8a',
+    'gray:40': '#adadad',
+    'gray:30': '#c9c9c9',
+    'gray:20': '#e0e0e0',
+    'gray:10': '#f2f2f2',
+    'orange:50': '#ff7633',
+    'yellow:50': '#faba2a',
+    'white:emphasis': 'rgba(255, 255, 255, 1.0)',
+    'white:primary': 'rgba(255, 255, 255, .92)',
+    'white:secondary': 'rgba(255, 255, 255, .60)',
+    'white:tertiary': 'rgba(255, 255, 255, .47)',
+    'white:disabled': 'rgba(255, 255, 255, .28)',
+    'black:emphasis': 'rgba(0, 0, 0, 1.0)',
+    'black:primary': 'rgba(0, 0, 0, .92)',
+    'black:secondary': 'rgba(0, 0, 0, .65)',
+    'black:tertiary': 'rgba(0, 0, 0, .54)',
+    'black:disabled': 'rgba(0, 0, 0, .30)',
+  });
+
+  expect(fonts).toEqual({
+    base: '"Segoe UI", -apple-system, BlinkMacSystemFont, "Helvetica Neue", Helvetica, Arial, sans-serif',
+    mono: '"Segoe UI Mono", "SFMono-Medium", "SF Mono", Menlo, Consolas, Courier, monospace',
+  });
+
+  expect(fontSizes).toEqual({
+    xs: '12px',
+    sm: '14px',
+    md: '16px',
+    lg: '18px',
+    xl: '20px',
+    '2xl': '24px',
+    '3xl': '28px',
+    '4xl': '32px',
+  });
+
+  expect(fontWeights).toEqual({
+    thin: 100,
+    extralight: 200,
+    light: 300,
+    normal: 400,
+    medium: 500,
+    semibold: 600,
+    bold: 700,
+    extrabold: 800,
+    black: 900,
+  });
+
+  expect(letterSpacings).toEqual({});
+
+  expect(lineHeights).toEqual({
+    normal: 'normal',
+    base: '1.5',
+    xs: '18px',
+    sm: '20px',
+    md: '22px',
+    lg: '24px',
+    xl: '28px',
+    '2xl': '32px',
+    '3xl': '36px',
+    '4xl': '40px',
+  });
+
+  expect(outlines).toEqual({
+    1: '1px solid',
+    2: '2px solid',
+    3: '3px solid',
+    4: '4px solid',
+    5: '5px solid',
+  });
+
+  expect(radii).toEqual({
+    circle: '50%',
+    none: 0,
+    sm: '3px',
+    md: '6px',
+    lg: '12px',
+  });
+
+  expect(shadows).toEqual({
+    none: 'none',
+  });
+
+  expect(sizes).toEqual(space);
+
+  expect(space).toEqual({
+    '1q': '1px',
+    '2q': '2px',
+    '3q': '3px',
+    '4q': '4px',
+    '5q': '5px',
+    '6q': '6px',
+    '7q': '7px',
+    '8q': '8px',
+    '9q': '9px',
+    '10q': '10px',
+    '1h': '2px',
+    '2h': '4px',
+    '3h': '6px',
+    '4h': '8px',
+    '5h': '10px',
+    '6h': '12px',
+    '7h': '14px',
+    '8h': '16px',
+    '9h': '18px',
+    '10h': '20px',
+    '1x': '4px',
+    '2x': '8px',
+    '3x': '12px',
+    '4x': '16px',
+    '5x': '20px',
+    '6x': '24px',
+    '7x': '28px',
+    '8x': '32px',
+    '9x': '36px',
+    '10x': '40px',
+    '11x': '44px',
+    '12x': '48px',
+    '13x': '52px',
+    '14x': '56px',
+    '15x': '60px',
+    '16x': '64px',
+    '17x': '68px',
+    '18x': '72px',
+    '19x': '76px',
+    '20x': '80px',
+    '24x': '96px',
+    '32x': '128px',
+    '40x': '160px',
+    '48x': '192px',
+    '56x': '224px',
+    '64x': '256px',
+  });
+
+  expect(zIndices).toEqual({
+    hide: -1,
+    auto: 'auto',
+    base: 0,
+    dropdown: 1000,
+    sticky: 1100,
+    fixed: 1200,
+    overlay: 1300,
+    drawer: 1400,
+    modal: 1500,
+    popover: 1600,
+    toast: 1700,
+    tooltip: 1800,
+  });
+});
 
 test('relative length units: rem', () => {
   const {
@@ -16,7 +277,7 @@ test('relative length units: rem', () => {
     sizes,
     space,
     zIndices,
-  } = rem;
+  } = createTheme('rem');
 
   expect(borders).toEqual({
     1: '.0625rem solid',
