+++ packages/theme/src/__tests__/index.test.js
@@ -1,15 +1,22 @@
 import defaultExport, * as moduleExport from '..';
+import { createTheme } from '..';
 
 test('should match expected exports', () => {
   const receivedExports = Object.keys(moduleExport);
   const expectedExports = [
     'default',
+
+    // createTheme
+    'createTheme',
   ];
 
   expect(receivedExports.sort()).toEqual(expectedExports.sort());
 });
 
 test('the default export must have all properties defined in the theme object', () => {
+  const theme = createTheme('rem');
+  expect(theme).toEqual(defaultExport);
+
   expect(defaultExport).toHaveProperty('borders');
   expect(defaultExport).toHaveProperty('breakpoints');
   expect(defaultExport).toHaveProperty('colors');
